refactor(db): align shared column builders
