+++ packages/db/src/shared/columns/build-image-url-column.ts
