+++ packages/db/src/shared/columns/build-pokemon-types-column.ts
