+++ packages/db/src/shared/columns/build-rarity-column.ts
@@ -1,0 +1,7 @@
+import { rarityEnum } from '../enums/rarity-enum.js';
+
+export function buildRarityColumn() {
+  return {
+    rarity: rarityEnum('rarity').notNull().default('D'),
+  };
+}
