--- packages/db/src/shared/columns/rarity-column.ts
@@ -1,7 +1,0 @@
-import { rarity } from '../enums/rarity-enum.js';
-
-export function buildRarityColumn() {
-  return {
-    rarity: rarity('rarity').notNull().default('D'),
-  };
-}
