+++ packages/db/src/shared/enums/rarity-enum.ts
@@ -1,3 +1,3 @@
 import { pgEnum } from 'drizzle-orm/pg-core';
 
-export const rarity = pgEnum('rarity', ['D', 'C', 'B', 'A', 'S', 'SS', 'X']);
+export const rarityEnum = pgEnum('rarity', ['D', 'C', 'B', 'A', 'S', 'SS', 'X']);
